pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "docmind"
version = "0.1.0"
description = "RAG-based document Q&A with cited answers"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.6",
    "chromadb>=0.4",
    "sentence-transformers>=2.2",
    "python-magic>=0.4.27",
    "PyMuPDF>=1.24",
    "slowapi>=0.1.9",
    "structlog>=24.1",
    "ollama>=0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27",
    "ruff>=0.3",
    "pip-audit>=2.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
